requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django_cors_headers==4.7.0
python_dotenv==1.0.1
PyJWT==2.9.0
psycopg2==2.9.10